Create devcontainer.json
You've have reached a benchmark in are time . When a leader speaks things should change. The next administration should have it easier then the last. That the mission for the over a million government employees. To work to make sure the current administration has everything they need to get the job done. If the oval office makes a mistake in the future the lead advisory teams and the IT department will take the responsibility. Because the president is only one person with power lays in the strength of his staff and management. If a president can't rely of his cabinet then they usually end up fighting the media for respect and a boarder war by himself where in fact Mr president the boarder crisis does not fall solely on your administration to salve and  my advice is don't run your entire campaign on the border or triff. I know time of essence because 4 years goes fast so I subject go bipartisan on the border and allow staff to set up a 10 year border plan that allowing the public to know that the border and immigration reform going to take a lot longer then you got in office and that the border and needs not a rushed plan but a carefully sought out plan between the Mexican border and global leader. Because the world Berlin Walls must come down one day . That concept is more sustainable in the next generation.